Move exporting to a file to a standalone function
