+++ sampledata/amsterdam-yymmdd.csv
@@ -1,0 +1,11 @@
+date,time,y,x,altitude
+2016/05/15,20:17:56,52.364788,4.871553,47.985538
+2016/05/15,20:17:57,52.364784,4.871575,41.04301099999999
+2016/05/15,20:17:58,52.364762,4.871584,36.924389000000005
+2016/05/15,20:17:59,52.364751,4.871588,36.544594
+2016/05/15,20:18:00,52.364748,4.871586,32.00935
+2016/05/15,20:18:01,52.364741,4.871573,31.454838000000002
+2016/05/15,20:18:02,52.364734,4.871565,31.809672999999997
+2016/05/15,20:18:03,52.364725,4.871557,33.017502
+2016/05/15,20:18:04,52.364719,4.871530,46.84192299999999
+2016/05/15,20:18:05,52.364705,4.871525,49.00122499999999
